test(factory): split LocalAI characterization fixtures
